requirements.txt
# Core ML / NLP
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.27.0

# Quantisation (optional, needed for --load_in_8bit / --load_in_4bit)
bitsandbytes>=0.42.0

# OpenAI API
openai>=1.20.0

# Data / numerics
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Plotting
matplotlib>=3.7.0
seaborn>=0.13.0

# Utilities
scikit-learn>=1.3.0
tqdm>=4.66.0
python-dotenv>=1.0.0
